pyproject.toml
[project]
name = "count-objects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "boto3>=1.37.5",
    "python-dotenv>=1.0.1",
]
